Get rid of BasePlayer class
